feat: :sparkles: ThemeToggle component
